requirements.txt
discord.py
discord.py[voice]
youtube_dl
PyNaCl==1.4.0
beautifulsoup4
lxml
requests
asyncio
holotools